add representation test for bulk vis job
